pyproject.toml
[build-system]
requires = ["setuptools >= 72.0"]
build-backend = "setuptools.build_meta"

[project]
name = "time-tools"
dynamic = ["version"]
description = "time-tools - collection of tools to handle timestamps"
authors = [
  {name = "Sebastian Weigmann", email = "seb.weigmann@gmail.com"}
]
maintainers = [
  {name = "Sebastian Weigmann", email = "seb.weigmann@gmail.com"}
]
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Programming Language :: Python :: 3",
  "Environment :: Console"
]
dependencies = [
  "python-dateutil >= 2.9.0"
  #"ssdeep>=3.4.1"
  #"ssdeep@git+https://github.com/DinoTools/python-ssdeep#egg=3.4.1",
  #"ssdeeper@git+https://codeberg.org/DFIR/python-ssdeeper#egg=3.5.1"
]
requires-python = ">=3.8"

[tool.setuptools.dynamic]
version = {attr = "time_tools.common.version.progver"}

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-v"
testpaths = [
    "test"
]

[project.urls]
#Homepage = "https://example.com"
#Documentation = "https://readthedocs.org"
Repository = "https://codeberg.org/DFIR/time-tools.git"
#"Bug Tracker" = "https://github.com/me/spam/issues"
#Changelog = "https://github.com/sweigmann/offset-tools/blob/main/CHANGELOG.md"

[project.scripts]
ts_to_isoutc = "time_tools.ts_to_isoutc:main"
